Further work towards making everything work

+++ receivers/native/arm/peripherals/clock/clock-ll-stm32f/src/main/c++/clock-ll-stm32f/specific/Clock.hpp
@@ -29,7 +29,7 @@ public:
         return freq / targetFreqKhz;
     }
 
-    int set(uint32_t targetFreqKhz, SystemClock source);
+    int8_t set(uint32_t targetFreqKhz, SystemClock source);
 
     uint32_t getFreqKhz() {
         return freq;
